Build and style the header and hero sections
The sections have been stylised. However, I would like the image
and .hero-text div to be the same height. I might try to fix it
when I come back but if I can't find a way I'll just leave it as
it is, it looks good enough (not too responsive but whatever, it's
not the purpose of the exercise anyways).