Finish styling the Info section
The Info section has be stylised, although maybe some larger gap
between images would be convenient. I haven't decided yet so I'll
leave it for the next time. A thing to keep in mind is I've realised
I shouldn't be coupling properties from different elements just because
they're equal (in a DRY manner), because it's more difficult to debug
later and change if needed. For the following sections, write properties
grouped by element, even if it means repeating properties. Only couple
when there's several elements that are exactly the same.